BoringUtilsTapSpec: Demonstrate wiring through when/layer breakage.
Boring non-probes out from a when isn't simply accomplished
(no unconditional when), but should work into a when (or layer).

For probes this should work up/through/into blocks (when/layer/...),
as well as across "sibling" blocks (e.g., out of one when, into its else)
which requires inserting a wire or port to bounce through.

May be rare to instantiate modules under `when`, but desiring
to bore/tap into (or out of) a layer is likely.